// frontend-new/pigikala/src/app/lib/types/api/responses/product-list-responses.ts
import { PaginatedResponse } from "./general-responses";

/**
 * 
 */
export interface ProductListItemResponse {
  id: number;
  name: string;
  mainPrice: number;
  mainImage: string | null;
  seller: SellerResponse;
}

export interface SellerResponse {
  id: number;
  storeName: string;
  storeImage: string | null;
  storeUrl: string;
}

/**
 * the name of subcategories are not unique, resort to id rather than name.
 */
export interface SubCategoryResponse {
  id: number;
  name: string;
}

export interface BreadCrumbResponse {
  subCategory: string;
  category: string
  mainCategory: string
}

export type ProductListResponse = PaginatedResponse<ProductListItemResponse>;